+++ app/Providers/AppServiceProvider.php
@@ -7,9 +7,11 @@ namespace App\Providers;
 
 use App\Models\Person;
 use App\Observers\PersonObserver;
+use App\Policies\PermissionManagementPolicy;
 use App\Policies\RoleManagementPolicy;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
+use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class AppServiceProvider extends ServiceProvider
@@ -31,5 +33,8 @@ class AppServiceProvider extends ServiceProvider
 
         // Register policy for Spatie Role model
         Gate::policy(Role::class, RoleManagementPolicy::class);
+
+        // Register policy for Spatie Permission model
+        Gate::policy(Permission::class, PermissionManagementPolicy::class);
     }
 }